% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_getStandardAnswer.R
\name{getAnswerByQuestion_dict}
\alias{getAnswerByQuestion_dict}
\title{GET}
\usage{
getAnswerByQuestion_dict(input, dict)
}
\arguments{
\item{input}{INPUT}

\item{dict}{DICT}
}
\value{
RETURN
}
\description{
GET
}
\examples{
getAnswerByQuestion_dict();
}
