% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_getStandardAnswer.R
\name{getStandardAsnwer_Manually}
\alias{getStandardAsnwer_Manually}
\title{从人工指定的地方获取相应的标准问答}
\usage{
getStandardAsnwer_Manually(brand = "JBLH")
}
\arguments{
\item{brand}{品牌}
}
\value{
返回值
}
\description{
从人工指定的地方获取相应的标准问答
}
\examples{
getStandardAsnwer_Manually();
}
