% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_CombineQuesByAnsw.R
\name{nsim_find_standardAnsw}
\alias{nsim_find_standardAnsw}
\title{find the answer}
\usage{
nsim_find_standardAnsw(data, dict)
}
\arguments{
\item{data}{data}

\item{dict}{dict}
}
\value{
return
}
\description{
find the answer
}
\examples{
nsim_find_standardAnsw()
}
