% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_CombineQuesByAnsw.R
\name{nsim_nscs_current}
\alias{nsim_nscs_current}
\title{获取最新版本的聚类数据}
\usage{
nsim_nscs_current(brand = "JBLH")
}
\arguments{
\item{brand}{品牌}
}
\value{
返回值
}
\description{
获取最新版本的聚类数据
}
\examples{
nsim_nscs_current();
}
