% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_CombineQuesByAnsw.R
\name{nsim_nscs_version}
\alias{nsim_nscs_version}
\title{问题聚类的函数}
\usage{
nsim_nscs_version(brand = "JBLH")
}
\arguments{
\item{brand}{品牌}
}
\value{
返回值
}
\description{
问题聚类的函数
}
\examples{
nsim_nscs_version();
}
