% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_CombineQuesByAnsw.R
\name{nsim_nscs_version_step1}
\alias{nsim_nscs_version_step1}
\title{处理数据，分步处理}
\usage{
nsim_nscs_version_step1(brand = "JBLH")
}
\arguments{
\item{brand}{品牌}
}
\value{
返回值
}
\description{
处理数据，分步处理
}
\examples{
nsim_nscs_version_step1();
}
