% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_ques_categoried_query.R
\name{ques_categoried_query_current}
\alias{ques_categoried_query_current}
\title{问题分类查询}
\usage{
ques_categoried_query_current(brand = "JBLH")
}
\arguments{
\item{brand}{品牌}
}
\value{
返回值
}
\description{
问题分类查询
}
\examples{
ques_categoried_query_current();
}
