% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_QuesClassified.R
\name{ques_category_list_current}
\alias{ques_category_list_current}
\title{获取意图模块}
\usage{
ques_category_list_current(brand = "JBLH")
}
\arguments{
\item{brand}{品牌}
}
\value{
返回值
}
\description{
获取意图模块
}
\examples{
ques_category_list_current()
}
