% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_QuesClassified.R
\name{ques_category_query_version}
\alias{ques_category_query_version}
\title{按版本查询问题分类}
\usage{
ques_category_query_version(brand = "JBLH", var_version)
}
\arguments{
\item{brand}{品牌}

\item{var_version}{nscs版本}
}
\value{
返回值
}
\description{
按版本查询问题分类
}
\examples{
ques_category_query_version();
}
