% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_CombineQuesByAnsw.R
\name{ques_combine_ByAnsw_query_standard}
\alias{ques_combine_ByAnsw_query_standard}
\title{查询}
\usage{
ques_combine_ByAnsw_query_standard(brand = "JBLH", var_version = "V1")
}
\arguments{
\item{brand}{品牌}

\item{var_version}{提示版本处理}
}
\value{
返回值
}
\description{
查询
}
\examples{
ques_combine_ByAnsw_query_standard();
}
