% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_ques_multiA.R
\name{ques_invalidOptions}
\alias{ques_invalidOptions}
\title{增加辅助无效功能}
\usage{
ques_invalidOptions(data)
}
\arguments{
\item{data}{数据}
}
\value{
返回值
}
\description{
增加辅助无效功能
}
\examples{
ques_invalidOptions();
}
