% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_ques_multiA.R
\name{ques_multiA_current_all}
\alias{ques_multiA_current_all}
\title{查询所有}
\usage{
ques_multiA_current_all(brand = "JBLH")
}
\arguments{
\item{brand}{品牌}
}
\value{
返回值
}
\description{
查询所有
}
\examples{
ques_multiA_current_all()
}
