% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_ques_multiA.R
\name{ques_multiA_current_lessThan}
\alias{ques_multiA_current_lessThan}
\title{检查问题的数量}
\usage{
ques_multiA_current_lessThan(brand = "JBLH", count = 5)
}
\arguments{
\item{brand}{品牌}

\item{count}{数量}
}
\value{
返回值
}
\description{
检查问题的数量
}
\examples{
ques_multiA_current_lessThan();
}
