% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_ques_multiA.R
\name{ques_multiA_current_moreThan}
\alias{ques_multiA_current_moreThan}
\title{查询记录大于部分记录}
\usage{
ques_multiA_current_moreThan(brand = "JBLH", count = 5)
}
\arguments{
\item{brand}{品牌}

\item{count}{记录}
}
\value{
返回值
}
\description{
查询记录大于部分记录
}
\examples{
ques_multiA_current_moreThan();
}
