% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_ques_multiA.R
\name{ques_standard_shedule}
\alias{ques_standard_shedule}
\title{查看问题的处理进度}
\usage{
ques_standard_shedule(brand = "JBLH")
}
\arguments{
\item{brand}{品牌}
}
\value{
返回值
}
\description{
查看问题的处理进度
}
\examples{
ques_standard_shedule();
}
