% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_ques_multiA.R
\name{ques_standard_top9_list}
\alias{ques_standard_top9_list}
\title{获取一个问题标准答案的数据}
\usage{
ques_standard_top9_list(brand = "JBLH")
}
\arguments{
\item{brand}{品牌}
}
\value{
返回值
}
\description{
获取一个问题标准答案的数据
}
\examples{
ques_standard_top9_list()
}
