% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_ques_categoried_query.R
\name{ques_uncategoried_query_current}
\alias{ques_uncategoried_query_current}
\title{未分类问题清单}
\usage{
ques_uncategoried_query_current(brand = "JBLH")
}
\arguments{
\item{brand}{品牌}
}
\value{
返回值
}
\description{
未分类问题清单
}
\examples{
ques_uncategoried_query_current();
}
